--- MobileAPI/README.md
@@ -1,2 +1,0 @@
-#技师端API
-1. [帐户及登录](1-1-account.md)
