--- MobileAPI/build.gradle
@@ -1,27 +1,0 @@
-group 'Autobon'
-version '1.0-SNAPSHOT'
-
-apply plugin: 'java'
-apply plugin: 'idea'
-apply plugin: 'org.flywaydb.flyway'
-
-repositories {
-    mavenCentral()
-}
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-dependencies {
-    compile "org.springframework.boot:spring-boot-starter:1.+"
-    compile "org.springframework.boot:spring-boot-starter-web:1.+"
-    compile "org.springframework.boot:spring-boot-starter-data-jpa:1.+"
-    compile 'org.springframework.boot:spring-boot-starter-security:1.+'
-    //compile 'org.springframework.security:spring-security-test:4.+'
-    compile "org.springframework.data:spring-data-redis:1.+"
-    compile 'redis.clients:jedis:2.+'
-    testCompile "org.springframework.boot:spring-boot-starter-test:1.+"
-    testCompile "com.jayway.jsonpath:json-path:2.+"
-    testCompile "org.mockito:mockito-core:1.+"
-    runtime "mysql:mysql-connector-java:5.+"
-}
