--- MobileAPI/src/main/java/com/autobon/mobile/repository/TechnicianRepository.java
@@ -1,13 +1,0 @@
-package com.autobon.mobile.repository;
-
-import com.autobon.mobile.entity.Technician;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * Created by dave on 16/2/13.
- */
-@Repository
-public interface TechnicianRepository extends JpaRepository<Technician, Integer> {
-    Technician getByPhone(String phone);
-}
