--- MobileAPI/src/main/java/com/autobon/mobile/security/TokenAuthenticationProvider.java
@@ -1,20 +1,0 @@
-package com.autobon.mobile.security;
-
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-
-/**
- * Created by dave on 16/2/17.
- */
-public class TokenAuthenticationProvider implements AuthenticationProvider {
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        return authentication;
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return TokenAuthentication.class.isAssignableFrom(authentication);
-    }
-}
