--- MobileAPI/src/main/java/com/autobon/mobile/utils/VerifyCode.java
@@ -1,183 +1,0 @@
-package com.autobon.mobile.utils;
-
-import javax.imageio.ImageIO;
-import java.awt.*;
-import java.awt.geom.AffineTransform;
-import java.awt.image.BufferedImage;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.util.Arrays;
-import java.util.Random;
-import java.util.stream.Collectors;
-
-/**
- * Created by dave on 16/2/14.
- * http://blog.csdn.net/ruixue0117/article/details/22829557
- */
-public class VerifyCode {
-    //验证码只显示大写，去掉了1,0,i,o几个容易混淆的字符
-    public static final String VERIFY_CODES = "23456789ABCDEFGHJKLMNPQRSTUVWXYZ";
-
-    public static String generateVerifyCode(int length) {
-        int codesLength = VERIFY_CODES.length();
-        Random random = new Random();
-        StringBuilder verifyCode = new StringBuilder(length);
-        for (int i = 0; i < length; i++) {
-            verifyCode.append(VERIFY_CODES.charAt(random.nextInt(codesLength - 1)));
-        }
-        return verifyCode.toString();
-    }
-
-    public static String generateRandomNumber(int length) {
-        Random random = new Random();
-        StringBuilder sb = new StringBuilder(length);
-        for (int i = 0; i < length; i++) {
-            sb.append(new Integer(random.nextInt(10)).toString());
-        }
-        return sb.toString();
-    }
-
-    /**
-     * 输出指定验证码图片流
-     * @param w
-     * @param h
-     * @param os
-     * @param code
-     * @throws IOException
-     */
-    public static void writeVerifyCodeImage(int w, int h, OutputStream os, String code) throws IOException {
-        int verifySize = code.length();
-        BufferedImage image = new BufferedImage(w, h, BufferedImage.TYPE_INT_RGB);
-        Random rand = new Random();
-        Graphics2D g2 = image.createGraphics();
-        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING,RenderingHints.VALUE_ANTIALIAS_ON);
-        Color[] colors = new Color[5];
-        Color[] colorSpaces = new Color[] { Color.WHITE, Color.CYAN,
-                Color.GRAY, Color.LIGHT_GRAY, Color.MAGENTA, Color.ORANGE,
-                Color.PINK, Color.YELLOW };
-        float[] fractions = new float[colors.length];
-        for(int i = 0; i < colors.length; i++){
-            colors[i] = colorSpaces[rand.nextInt(colorSpaces.length)];
-            fractions[i] = rand.nextFloat();
-        }
-        Arrays.sort(fractions);
-
-        g2.setColor(Color.GRAY);// 设置边框色
-        g2.fillRect(0, 0, w, h);
-
-        Color c = getRandColor(200, 250);
-        g2.setColor(c);// 设置背景色
-        g2.fillRect(0, 2, w, h-4);
-
-        //绘制干扰线
-        Random random = new Random();
-        g2.setColor(getRandColor(160, 200));// 设置线条的颜色
-        for (int i = 0; i < 20; i++) {
-            int x = random.nextInt(w - 1);
-            int y = random.nextInt(h - 1);
-            int xl = random.nextInt(6) + 1;
-            int yl = random.nextInt(12) + 1;
-            g2.drawLine(x, y, x + xl + 40, y + yl + 20);
-        }
-
-        // 添加噪点
-        float yawpRate = 0.05f;// 噪声率
-        int area = (int) (yawpRate * w * h);
-        for (int i = 0; i < area; i++) {
-            int x = random.nextInt(w);
-            int y = random.nextInt(h);
-            int rgb = getRandomIntColor();
-            image.setRGB(x, y, rgb);
-        }
-
-        shear(g2, w, h, c);// 使图片扭曲
-
-        g2.setColor(getRandColor(100, 160));
-        int fontSize = h-4;
-//        Font font = new Font("Algerian", Font.ITALIC, fontSize);
-//        g2.setFont(font);
-        char[] chars = code.toCharArray();
-        for(int i = 0; i < verifySize; i++){
-            AffineTransform affine = new AffineTransform();
-            affine.setToRotation(Math.PI / 4 * rand.nextDouble() * (rand.nextBoolean() ? 1 : -1), (w / verifySize) * i + fontSize/2, h/2);
-            g2.setTransform(affine);
-            g2.drawChars(chars, i, 1, ((w-10) / verifySize) * i + 5, h/2 + fontSize/2 - 10);
-        }
-
-        g2.dispose();
-        ImageIO.write(image, "jpg", os);
-    }
-
-    private static Color getRandColor(int foreColor, int backColor) {
-        Random random = new Random();
-        if (foreColor > 255)
-            foreColor = 255;
-        if (backColor > 255)
-            backColor = 255;
-        int r = foreColor + random.nextInt(backColor - foreColor);
-        int g = foreColor + random.nextInt(backColor - foreColor);
-        int b = foreColor + random.nextInt(backColor - foreColor);
-        return new Color(r, g, b);
-    }
-
-    private static int getRandomIntColor() {
-        Random random = new Random();
-        int color = 0;
-        for (int i = 0; i < 3; i++) {
-            color = color << 8;
-            color = color | random.nextInt(255);
-        }
-        return color;
-    }
-
-    private static void shear(Graphics g, int w1, int h1, Color color) {
-        shearX(g, w1, h1, color);
-        shearY(g, w1, h1, color);
-    }
-
-    private static void shearX(Graphics g, int w1, int h1, Color color) {
-        Random random = new Random();
-        int period = random.nextInt(2);
-
-        boolean borderGap = true;
-        int frames = 1;
-        int phase = random.nextInt(2);
-
-        for (int i = 0; i < h1; i++) {
-            double d = (double) (period >> 1)
-                    * Math.sin((double) i / (double) period
-                    + (6.2831853071795862D * (double) phase)
-                    / (double) frames);
-            g.copyArea(0, i, w1, 1, (int) d, 0);
-            if (borderGap) {
-                g.setColor(color);
-                g.drawLine((int) d, i, 0, i);
-                g.drawLine((int) d + w1, i, w1, i);
-            }
-        }
-
-    }
-
-    private static void shearY(Graphics g, int w1, int h1, Color color) {
-        Random random = new Random();
-        int period = random.nextInt(40) + 10; // 50;
-
-        boolean borderGap = true;
-        int frames = 20;
-        int phase = 7;
-        for (int i = 0; i < w1; i++) {
-            double d = (double) (period >> 1)
-                    * Math.sin((double) i / (double) period
-                    + (6.2831853071795862D * (double) phase)
-                    / (double) frames);
-            g.copyArea(i, 0, 1, h1, 0, (int) d);
-            if (borderGap) {
-                g.setColor(color);
-                g.drawLine(i, (int) d, i, 0);
-                g.drawLine(i, (int) d + h1, i, h1);
-            }
-
-        }
-
-    }
-}
