--- MobileAPI/src/main/resources/application.yml
@@ -1,36 +1,0 @@
-com.autobon:
-    sms.gateway: http://114.215.172.92/message/messageSend # SMS Gateway
-    env: TEST
-
-# Tomcat
-server.port: 51234
-
-# Logging level (TRACE, DEBUG, INFO, WARN, ERROR, FATAL, OFF)
-logging:
-    level:
-        org.hibernate: WARN
-        org.springframework: WARN
-    file: Autobon.log
-
-# spring
-spring:
-    profiles.active: dev
-    freemarker.cache: false
-    resources.cache-period: 7200
-    datasource:
-        url: jdbc:mysql://127.0.0.1:3306/autobon?characterEncoding=utf8&useSSL=false
-        username: root
-        password: 123456
-    jpa:
-        show-sql: true
-        hibernate:
-            # Hibernate DDL  auto: create, create-drop, update, validate
-            ddl-auto: validate
-            naming-strategy: org.hibernate.cfg.ImprovedNamingStrategy
-            properties:
-                hibernate.dialect: org.hibernate.dialect.MySQL5Dialect
-    redis:
-        host: 120.55.98.235
-        port: 6379
-        password: 193ete0AD2936ac06c09&5307fab776Zcd0fatkf0c4ef
-
