+++ README.md
@@ -1,5 +1,2 @@
-# Platform
-车邻邦平台端
-
-##API文档
-[技师端API](MobileAPI/README.md)
+#技师端API
+1. [帐户及登录](1-1-account.md)
