+++ Technician/build.gradle
@@ -1,14 +1,32 @@
 group 'Autobon'
 version '1.0-SNAPSHOT'
 
-apply plugin: 'java'
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:1.+")
+        classpath("org.flywaydb:flyway-gradle-plugin:3.+")
+    }
+}
 
-sourceCompatibility = 1.5
+apply plugin: 'java'
+apply plugin: 'idea'
+apply plugin: 'spring-boot'
+apply plugin: 'org.flywaydb.flyway'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.11'
+    compile "org.springframework.boot:spring-boot-starter:1.+"
+    compile 'org.springframework.boot:spring-boot-starter-security:1.+'
+    compile "org.springframework.boot:spring-boot-starter-data-jpa:1.+"
+    testCompile "org.springframework.boot:spring-boot-starter-test:1.+"
+    runtime "mysql:mysql-connector-java:5.+"
 }
