+++ Technician/src/main/java/com/autobon/technician/Application.java
@@ -1,0 +1,18 @@
+package com.autobon.technician;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.orm.jpa.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@SpringBootApplication(scanBasePackages = {"com.autobon"})
+@EntityScan(basePackages = {"com.autobon"})
+@EnableJpaRepositories(basePackages = {"com.autobon"})
+@EnableTransactionManagement
+public class Application {
+    public static void main(String[] args){
+        SpringApplication.run(Application.class, args);
+    }
+}
