+++ Technician/src/main/resources/application.yml
@@ -1,6 +1,3 @@
-com.autobon:
-    sms.gateway: http://114.215.172.92/message/messageSend # SMS Gateway
-    env: PROD
 
 # Tomcat
 server.port: 51234
@@ -18,9 +15,9 @@ spring:
     freemarker.cache: true
     resources.cache-period: 7200
     datasource:
-        url: jdbc:mysql://127.0.0.1:3306/autobondb?characterEncoding=utf8&useSSL=false
-        username: autobon
-        password: autobon
+        url: jdbc:mysql://127.0.0.1:3306/autobon?characterEncoding=utf8&useSSL=false
+        username: root
+        password: 123456
     jpa:
         show-sql: true
         hibernate:
@@ -29,7 +26,3 @@ spring:
             naming-strategy: org.hibernate.cfg.ImprovedNamingStrategy
             properties:
                 hibernate.dialect: org.hibernate.dialect.MySQL5Dialect
-    redis:
-        host: 120.55.98.235
-        port: 6379
-        password: 193ete0AD2936ac06c09&5307fab776Zcd0fatkf0c4ef
