+++ build.gradle
@@ -24,12 +24,17 @@ sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
 dependencies {
-    compile project(":MobileAPI")
-    compile("org.springframework.boot:spring-boot-starter-data-jpa:1.+")
-    compile("org.springframework.boot:spring-boot-starter-freemarker:1.+")
-    testCompile("org.mockito:mockito-core:1.+")
-    testCompile("org.springframework.boot:spring-boot-starter-test:1.+")
-    runtime("mysql:mysql-connector-java:5.+")
+    compile project(":Technician")
+    compile "org.springframework.boot:spring-boot-starter:1.+"
+    compile "org.springframework.boot:spring-boot-starter-web:1.+"
+    compile 'org.springframework.boot:spring-boot-starter-security:1.+'
+    compile "org.springframework.boot:spring-boot-starter-data-jpa:1.+"
+    compile "org.springframework.boot:spring-boot-starter-freemarker:1.+"
+    compile "org.springframework.data:spring-data-redis:1.+"
+    compile 'redis.clients:jedis:2.+'
+    testCompile "org.springframework.boot:spring-boot-starter-test:1.+"
+    testCompile "com.jayway.jsonpath:json-path:2.+"
+    runtime "mysql:mysql-connector-java:5.+"
 }
 
 gradle.projectsEvaluated {
