+++ settings.gradle
@@ -1,5 +1,2 @@
 rootProject.name = 'Platform'
-include 'MobileAPI'
 include 'Technician'
-include 'Shared'
-
