--- src/main/java/com/autobon/mobile/controller/Greeting.java
@@ -1,23 +1,0 @@
-package com.autobon.mobile.controller;
-
-/**
- * Created by liz on 2016/1/27.
- */
-public class Greeting {
-
-    private final long id;
-    private final String content;
-
-    public Greeting(long id, String content) {
-        this.id = id;
-        this.content = content;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getContent() {
-        return content;
-    }
-}
