+++ src/main/java/com/autobon/platform/controller/Greeting.java
@@ -1,0 +1,23 @@
+package com.autobon.platform.controller;
+
+/**
+ * Created by liz on 2016/1/27.
+ */
+public class Greeting {
+
+    private final long id;
+    private final String content;
+
+    public Greeting(long id, String content) {
+        this.id = id;
+        this.content = content;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
